+++ app/src/main/java/com/example/rolloprint/PrintServerService.kt
@@ -8,6 +8,7 @@ import android.app.Service
 import android.content.Context
 import android.content.Intent
 import android.content.pm.ServiceInfo
+import android.graphics.Bitmap
 import android.net.nsd.NsdManager
 import android.net.nsd.NsdServiceInfo
 import android.net.wifi.WifiManager
@@ -78,7 +79,12 @@ class PrintServerService : Service() {
         acquirePowerLocks()
     }
 
-    fun initializeServer(usbPrintManager: UsbPrintManager, logger: (String) -> Unit, onStatusChanged: (Boolean, String?) -> Unit) {
+    fun initializeServer(
+        usbPrintManager: UsbPrintManager,
+        logger: (String) -> Unit,
+        onStatusChanged: (Boolean, String?) -> Unit,
+        onNetworkBitmapRendered: ((Bitmap) -> Unit)? = null
+    ) {
         if (ippServer == null) {
             ippServer = IppServer(this, usbPrintManager, { logMsg ->
                 logger(logMsg)
@@ -86,7 +92,7 @@ class PrintServerService : Service() {
                 isServerRunning = running
                 updateNotification(if (running) "Active on $ip:$PORT" else "Server Stopped")
                 onStatusChanged(running, ip)
-            })
+            }, onNetworkBitmapRendered)
         }
         ippServer?.start()
         registerNsdService(logger)